rename AbstractChallengeResponseAction
